requirements.txt
# Core document processing
PyMuPDF>=1.23.0  # fitz - fast PDF processing
pdfplumber>=0.10.0  # accurate table extraction
pytesseract>=0.3.10  # OCR for formulas/images
Pillow>=10.0.0  # image processing for OCR
pypdf>=3.17.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# API and web framework
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# RAG and LLM
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
openai>=1.3.0
instructor>=0.4.0

# Vector databases
pinecone-client>=3.0.0
pinecone>=3.0.0

# Embeddings
sentence-transformers>=2.2.0
tiktoken>=0.5.0  # token counting

# Wikipedia fallback
wikipedia>=1.4.0

# Airflow
apache-airflow>=2.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
tqdm>=4.64.0

